move away WorkerThread from Main class
