+++ Lesson5/Lesson5.iml
@@ -9,4 +9,3 @@
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
 </module>
-
